Move workspace logic to appservice
